refactor: Human 상속 받는 객체에 공통 메서드 선언 위치 이동
