+++ README.md
@@ -1,10 +1,14 @@
-# java blackjack
+# java-blackjack
+
+### 로또 미션 저장소
 
 블랙잭 게임은 딜러와 플레이어 중 카드의 합이 21 또는 21에 가장 가까운 숫자를 가지는 쪽이 이기는 게임이다.
 
-## 기능 구현 목록
+## ♠ 구현해야 하는 기능
 
-### 플레이어 등록
+> 입력 예외 발생 시 `Illegalargumentexception` 을 발생시킨다.
+
+### 📌 플레이어 등록
 - [x] 게임에 참여할 사람의 이름을 입력받는다.
   - [예외처리] 이름은 빈 값을 입력하면 안된다.
   - [예외처리] 이름은 null 을 입력하면 안된다.
@@ -12,7 +16,7 @@
   - [예외처리] 이름은 `딜러`와 같으면 안된다.
   - [예외처리] 게임에 참여하는 사람은 딜러 포함 26명을 넘길 수 없다.
 
-### 카드 뽑기
+### 📌 카드 뽑기
 - [x] 게임에 참여할 사람의 목록을 입력 받으면 모든 종류의 카드를 생성한다.
 - [x] 생성한 카드를 랜덤으로 섞는다.
 - [x] 카드를 나누어줄 때에는 랜덤으로 섞은 카드를 뒤에서부터 하나씩 준다.
@@ -26,13 +30,11 @@
   - [예외처리] `y`, `n` 이외의 값을 입력하면 안된다.
 - [x] 모든 참가자를 확인한 후 딜러의 총 합이 `17` 점을 넘길 때 까지 계속해서 카드를 추가한다.
 
-### 승패 결정
+### 📌 승패 결정
 - [x] 딜러의 턴이 종료되면 가지고 있는 모든 카드와 총 합을 오픈한다.
 - [x] 최종 승패를 결정한다.
 
-> 입력 예외 발생 시 `Illegalargumentexception` 을 발생시킨다.
-
-### 실행 결과
+## 실행 결과
   ```
     게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)
     pobi,jason
